Add `./Build bundle` command.
It uses Carton to install all prereqs into a single directory, then
installs Sqitch into that directory and adds an appropriate `use lib`
line to the Sqitch app installed there. It shoudl be that Sqitch canb
be run from that directory without any other Perl dependencies other
tha Perl itself.

The hope is to be able to more easily create simple bundles for
distribution in environments such as Homebrew, a Docker image, or an
OS-specific binary installer.