Use Menlo::CLI::Compat and t/sqitch for bundle.
The former is the code behind both Carton and cpanm, so just take
advntage of it.  As to using t/sqitch, it uses FindBin, so we can
avoid parsing it to insert a `use lib`. But I had to switch from
`/usr/bin/env` to just `#!perl` to get Module::Build's shebang fix
code to work.

Also, make an explicit `cpanfile` to maintain for this purpose. Bit of
a PITA to have yet another list of dependencies to maintain, but it
simplifies things a bit. Sadly, the `--with-recommends` option seems
to be ignored, or perhaps it isn't relying on the `cpanfile` for
requires and recommends? Will have to figure that out.

And finally, add some code to remvoe files not necessary for the bundle,
including packlists, metadata, and build and test modules.